set up basic menu- moved everythng but HOME into components
