+++ client/src/App.js
@@ -2,12 +2,14 @@ import React from 'react';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import { ApolloProvider } from '@apollo/react-hooks';
 import ApolloClient from 'apollo-boost';
-
+//Navbar is the header - note I couldn't import jsx?
 import Navbar from './components/Header/Navbar';
 import Footer from './components/Footer/Footer';
-import About from './pages/About';
+import About from './components/About/About';
 import Home from './pages/Home';
-import Login from './pages/Login';
+import Login from './components/LoginForms/LoginForm';
+import PortfolioBuilder from './components/PortfolioBuilder/PortfolioBuilder';
+
 
 const client = new ApolloClient({
   request: (operation) => {
@@ -32,6 +34,8 @@ const client = new ApolloClient({
               <Route exact path='/' component={Home} />
               <Route exact path='/about' component={About} />
               <Route exact path='/login' component={Login} />
+              <Route exact path='/portfolio' component={PortfolioBuilder} />
+
               <Route render={() => <h1 className='display-2'>Wrong page!</h1>} />
             </Switch>
             <Footer />
