+++ client/src/components/About/About.js
@@ -7,11 +7,11 @@ const About = () => {
   return (
     <div>
       <div className="flex-row justify-space-between">
-        
+      <div className="container">
         <h2>Test About Page</h2>
-        <h3>Might want a component for realsies</h3>
+        <p>This page currently does not require login</p>
         
-        
+        </div>
       </div>
     </div>
   );
