+++ client/src/components/LoginForms/LoginForm.js
@@ -71,7 +71,7 @@ const LoginForm = () => {
             value={userFormData.email}
             required
           />
-          <Form.Control.Feedback type='invalid'>Email is required!</Form.Control.Feedback>
+          <Form.Control.Feedback type='invalid'>Email is required</Form.Control.Feedback>
         </Form.Group>
 
         <Form.Group>
@@ -84,7 +84,7 @@ const LoginForm = () => {
             value={userFormData.password}
             required
           />
-          <Form.Control.Feedback type='invalid'>Password is required!</Form.Control.Feedback>
+          <Form.Control.Feedback type='invalid'>Password is required</Form.Control.Feedback>
         </Form.Group>
         <Button disabled={!(userFormData.email && userFormData.password)} type='submit' variant='success'>
           Submit
